+++ packages/ricos-theme/web/src/palettes/utils.ts
@@ -2,13 +2,10 @@ import { difference } from 'lodash';
 import { Palette } from 'ricos-common';
 
 const BG_COLOR = 11;
-const SECONDARY_COLOR = 13;
-const COLOR4 = 14;
 const TEXT_COLOR = 15;
-const COLOR7 = 17;
 const ACTION_COLOR = 18;
 
-export const COLORS = { BG_COLOR, SECONDARY_COLOR, COLOR4, TEXT_COLOR, COLOR7, ACTION_COLOR };
+export const COLORS = { BG_COLOR, TEXT_COLOR, ACTION_COLOR };
 export const assertPalette = (palette?: Palette) => {
   if (!palette) return;
   if (palette.length === 0) throw Error('Received empty palette colors array');