+++ CHANGELOG.md
@@ -16,11 +16,20 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+- `ricos-theme`
+  - [#1550](https://github.com/wix-incubator/rich-content/pull/1550) fixes:
+    - plugin menu color fix
+    - resize handles color fix
+    - brightness threshold to activate fallback color - raised
+
 ### 🏠 Internal
 - `text-selection-toolbar`
   - [#1412](https://github.com/wix-incubator/rich-content/pull/1412) text selection toolbar example in storybook
 - `storybook`
   - [#1412](https://github.com/wix-incubator/rich-content/pull/1412) text selection toolbar example
+- `ricos-theme`
+  - [#1550](https://github.com/wix-incubator/rich-content/pull/1550) removed unused-by-design colors from palette (secondary, color4, color7)
   
 </details>
 <hr/>